all changes before final setup
this is the code i have been using the whole time. commits after this will be code i use to get to final results with all outliers dealt with, all terminations =2, etc. so this commit is just 'backup' in case i need to find old code or run previous analyses again.